Updated the symbol-table test to match the rest of the tests

+++ trunk/tests/test-symbol-table.sh
@@ -1,3 +1,6 @@
+#!/bin/sh
+# This is meant to be run from the root of the losak source directory.
+echo "Testing the symbol-table support..."
 PROGRAM="(define (rreverse list result return)
   (cond ((atom list) (return result))
         ('t (rreverse (cdr list) (cons (car list) result) return))))
@@ -17,11 +20,18 @@ PROGRAM="(define (rreverse list result return)
                  (symbol->string (string->symbol \"World!\"))))"
 echo ${PROGRAM} | ./bin/symbol-table | ./bin/desugar | ./bin/compiler > main.c
 gcc *.c -o bin/test-symbol-table
-EXPECTED="Hello, World!"
-ACTUAL=$(./bin/test-symbol-table)
-if [[ ${ACTUAL} == ${EXPECTED} ]]; then
-    echo PASSED
+if [ $? ]; then
+  EXPECTED="Hello, World!"
+  ACTUAL=$(./bin/test-symbol-table)
+  if [ "$ACTUAL" = "$EXPECTED" ]; then
+    echo "\tPassed"
+    return 0
+  else
+    echo "\tFailed"
+    return 1
+  fi
+  rm bin/test-symbol-table
 else
-    echo FAILED ${ACTUAL}
+  echo "\tBuild Failed"
+  return 1
 fi
-rm bin/test-symbol-table
